[HACK] open5gs/ausf: Call dauth shim from AUSF auth request handler
This commit adds a call to the rpc shim from within the AUSF when it
receives an authentication request on the Nausf api endpoint. This
commit is hacky, since it calls the RPC from within the main context,
which previously was written to not block the main ausf server
thread.

The ausf handles each request by scheduling outgoing requests to the
UDM, and then only generating the reply to the original request later
when triggered by the received http/2 reply from the UDM. This makes
it difficult to add support for also sending requests and receiving
replies via GRPC, since then two different listeners could generate
concurrent events, and the ausf is currently explicilty single
threaded.

Instead of addressing the threading issues, this commid just makes a
synchronous rpc call from the request receive context, blocking AUSF
processing until it times out or receives a reply. As a first attempt
this should be sufficient to prove correctness, at the expense of
performance. We'll likely eventually one to fix this to allow the
dAuth manager to handle multiple requests at the same time.