open5gs/ausf: Refactor grpc dep specification to list format
Grpc contains some transitive dependencies that are not picked up by
the top-level targets. To make it more convenient to include them all,
this commit refactors the way the grpc deps are added to define a list
of them up front, and then include that list in the relevant
targets.