manager-rs, protos: Incorporate local_authentication into manager
In addition to adding the build and import code to the rust manager,
this commit changes the package of the local_authentication proto to
use snake_case instead of camelCase in the package name. Tonic and/or
prost was unable to build and import the package for some reason with
a camelCase name.