[protos] Split package namespace between local and remote
The local service is intended specifically for interfacing with a
local network core and supporting functions, while the remote service
needs to be externally facing. Since some of the names are similar,
split them into different package namespaces to help make sure the
messages are not used in incorrect parts of the code.