manager: Use strings for user id instead of u8 vectors
Strings are slightly less efficient since they are sometimes heap
allocated, but allow for more fleixbility in the id type and natively
support lots of the conversions and comparisons that would need to be
manually implemented otherwise. This is important for interfacing with
open5gs, since the SBI interface expects valid utf-8 ids of a variable
length rather than binary.