[open5gs] [HACK] Don't run tests in debian package build
Since we're not building the packages on different architectures or as
part of a formal process, it's not necessary to run the tests as part
of the build. This allows the build to happen on a machine where the
tests would fail, i.e. a machine that cannot create temporary sockets
at specific well-known ports.