[open5gs] Update open5gs package names for dauth
To make it easer to track what is deployed in the testing environment,
this commit updates the package names to include "dauth" in the name
so we can tell them apart from upstream versions more easily. All
services, binaries, directories, and config files are still their
regular names, this change only impacts the package names themselves.