[scripts] Update ip config script to pull data from file
When deploying new versions of the amf and upf, the configuration
files could be overwritten with new information and need to be
regenerated. This commit updates the ip config script to pull the ip
from the dauth/host-config.yml file (written when the system is
provisioned) when no IP is explicitly provided on the command line.