Expand confirm auth vector
Added new checks to determine where the auth vector originated from, so
that the kseaf could be correctly found. Added auth state to help with
this, which holds the rand value of a vector, the user it is for, and
whether it was from the home network or a backup network. Needs key
share functionality to be completed.