manager-rs: Integrate key share generation into dauth
This commit adds support for generating, propagating, and eventually
re-deriving kseaf from backup network key shares. It uses the
KseafShare type from the keys module to help give a strong typing
guarantee that the share key material is not accidentally mixed up
with the original key material.

The implementation currently uses a constant value for the number of
shares required to recover the key, although this could be set on a
per-user or per-network basis in the future.