auth-vector dauth: Convert Sqn to full type
The SQN was just a type alias to a vector of u8s, which doesn't allow
implementation of specific conversion traits. This commit changes the
Sqn to be an actual type, which allows conversion traits to be
specified for the SQN near its definition, and should allow for better
type checking and reliable conversion the values are passed around the
system. The SQN in particular is used a lot, so I think it make sense
to implement all the traits for it while other less uses fields can
probably remain type aliases for now.