Fix directory service user info bug
The directory service was only getting user upsert requests when the
user had at least one backup network. This is because the update user
task originally only kept entries for a given user and its backup
networks. There is now a special entry for the home networks that is
filtered out of the backups before being sent to the directory service.