services: Only wait for the minimum number of valid key responses
Rather than waiting for all backup networks to either respond or
timeout, this commit uses a JoinSet to join each response as it
becomes available (regardless of the order the responses return
in). Once a sufficient number of responses are received, the loop is
exited early to allow the serving network to go on to use the key as
quickly as possible.