services: Don't delete auth vectors until confirmed
This commit is a fairly significant change in the auth vector handling
logic. It changes the backup network hander to not delete a given
stored auth vector at the point it is originally sent to the serving
network, and instead only delete it once the confirmation is received
for that vector.

This allows the serving network to request multiple auth vectors from
multiple backup networks in parallel, but comes at the cost of
potentially losing synchronization with the UE if the serving network
does send this vector to the UE and the UE provides a RES, but it is
lost. This commit adds a "sent" column to the auth vector store, which
would allow a more error-tolerant implementation to detect the
synchronization error from the UE and then send a fresh vector that
had never been sent previously.