services: Re-work vector reporting to allow multiple reports
If we allow a vector to be used in multiple races, then it's possible
it can be sent multiple times before being used. In the long run the
best answer is probably to differentiate between knowing the vector
has been used vs. just knowing the vector participated in a race. For
now this commit re-works the reporting logic to be okay if multiple
uses are reported, and then to only return a new vector after the
first use.

This will mean in practice that when a home network comes back online
it might have to replace a lot more vectors than actually were truly
used, but this is probably okay since vectors are plentiful and the
maximum number of "extra" vectors is still just the number of backup
networks since the backup will keep re-racing the same vector until
it's actually used for keying.