[HACK] services/dauth: Always keep the greatest SQN
The previous behavior replaced the SQNs with what was in the config
file on every startup. This works okay for a test environment, but
means you need to update the config every time you do an
authentication with a real sim : ) In the long run the whole database
architecture could probably be cleaned up to use some kind of
migration/seeding pattern that would make this easier to manage. For
now this commit adds special case handling to the update of the SQN
max for each particular slice.