services/dauth: Semi-handle resynchronization information
While the contents of the resync data are not incredibly useful in the
dauth regime (since they only tell you the next available max SQN,
which might correspond to a different SQN vector), they do offer a
signal that the auth you just used was too old. This commit updates
the serving network to send a hint back to all of the backups on its
next request with the xres_star_hash of the auth used that caused the
resync. This allows the backups to advance the relevant SQN vector if
it belongs to them.

From a security perspective, this allows the serving network to
effectively consume an auth they know about without getting a valid
response from the UE. In a production implementation some limiting
would need to be added to the backup networks to prevent auth
exhaustion by a poorly behaving serving network.