services/dauth: Flatten backup update transactions per user
Transactions were being created within the inner loop of per each
backup per each vector per each user... this makes the initial user
update process very slow on nodes with slow disk. This commit
refactors this code to move the transaction to the outermost loop, but
at the expense of larger transaction chunks and bigger sections that
would have to be recovered in the event of a failure.